Code Review Changes
- Updated folder and project formatting to [ProjectName].[UserName]
- Added a README file
- Switched to File-Scoped namespaces
- Removed double casting in Menu.GetNumberInput()
- Added Menu.ValidateRecordId() and SqlHelper.GetAllIds to fix Menu.UpdateRecord() and stop the user from entering an Id that doesn't exist
