+++ HabitTracker.MichaelKarcz/HabitTrackerLibrary/Models/Habit.cs
@@ -1,0 +1,8 @@
+﻿namespace HabitTrackerLibrary.Models;
+public class Habit
+{
+    public int Id { get; set; }
+    public string? Date { get; set; }
+    public int Miles { get; set; }
+
+}
