--- HabitTracker/HabitTrackerLibrary/Models/Habit.cs
@@ -1,10 +1,0 @@
-﻿namespace HabitTrackerLibrary.Models
-{
-    public class Habit
-    {
-        public int Id { get; set; }
-        public string? Date { get; set; }
-        public int Miles { get; set; }
-
-    }
-}
